// server/config/interface.ts
import { Request } from 'express';
import { Document } from 'mongoose';

export interface IUser extends Document {
    name: string
    account: string
    password: string
    avatar: string
    role: string
    type: string
    rf_token?: string
    _doc: object
}

export interface INewUser {
    name: string
    account: string
    password: string
}

export interface IDecodeToken {
    id?: string
    newUser?: INewUser
    iat: number
    exp: number
}

export interface IGLPayload {
    email: string
    name: string
    email_verified: boolean
    picture: string
}

export interface IUserParams {
    account: string
    name: string
    avatar?: string
    password: string
    type: string
}

export interface IReqAuth extends Request {
    user?: IUser
}

export interface IComment extends Document {
    user: string
    blog_id: string
    blog_user_id: string
    content: string
    replyCM: string[]
    reply_user: string
    comment_root: string
    _doc: object
}